requirements_dev.txt
# Misc
requests
python-dotenv
classmods
persian

# Flask
flask
gunicorn

# API Connections
pynetbox
py-ami-client
py-sarvcrm-api

pytest